requirements.txt
APScheduler>=3.10.4
holidays>=0.57
openai>=1.40.0
pandas>=2.2.2
Pillow>=10.4.0
pytz>=2024.1
openpyxl>=3.1.5
Flask>=2.3.0
requests>=2.31.0
